# specs/001-shopping-voice-dashboard/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Life OS Dashboard API
  version: 0.1.0
  description: >
    Authenticated HTTP API for the Life OS web dashboard (MVP). All routes
    require a session cookie established via POST /auth/login unless noted.
servers:
  - url: https://api.example.com
    description: Production (placeholder)
  - url: http://localhost:3000
    description: Local NestJS
...
paths:
...
  /ideas:
    get:
      tags: [Ideas]
      security:
        - cookieAuth: []
      responses:
        "200":
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Idea"
    post:
      tags: [Ideas]
      security:
        - cookieAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [body]
              properties:
                body:
                  type: string
                  maxLength: 10000
                title:
                  type: string
                  maxLength: 200
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Idea"
  /ideas/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    delete:
      tags: [Ideas]
      security:
        - cookieAuth: []
      responses:
        "204":
          description: Deleted
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: lifeos_session
  schemas:
...
    Idea:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
          nullable: true
        body:
          type: string
        created_at:
          type: string
          format: date-time
...